sql: move tableDeleter to its own file
Release note: None
